+++ goDB/Implementations/Base.php
@@ -100,14 +100,16 @@ abstract class Base
      * Validates and normalizes connection parameters
      *
      * @param array $params
+     * @param string $error
      * @return array
      *         normalized parameters or FALSE if parameters is invalid
      */
-    public function checkParams(array $params)
+    public function checkParams(array $params, &$error = null)
     {
         $result = array();
         foreach ($this->paramsReq as $param) {
             if (!array_key_exists($param, $params)) {
+                $error = 'required '.$param;
                 return false;
             }
             $result[$param] = $params[$param];