+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * #44: Extended ?col & ?cols format
 * Fixed ?order (if scalar it does not add "ASC")
+* Added to travis-ci.org
+* ConfigError: normal message
 
 ## 2.2.6 (12.02.2016)